--- a/.eslintrc.json
+++ b/.eslintrc.json
@@ -111,7 +111,7 @@
     ],
     "complexity": [
       "error",
-      10
+      12
     ],
     "max-depth": [
       "error",