--- a/manual_test/specs/regex-pattern-parser-across-elements.js
+++ b/manual_test/specs/regex-pattern-parser-across-elements.js
@@ -7,15 +7,15 @@ describe('test the \'collectRegexGroupIndexes\' method', function() {
     reg = /(?<=f)(a(?:a(a))a(a))((?<gr5>\2)+)(b(b(?<n>b))(?:bb)(?!a))()(?:)((?:\k<n>|\(|\)|\\)+)([a-z/()[\]\\]+?)(?=d)/;
 
   beforeEach(function() {
-    $ctx = $('.regex-pattern-parser');
-    $ctx.unmark();
+    $ctx = $('.regex-pattern-parser-across-elements');
   });
 
-  // test 'separateGroups' method
-  it('should count and test content of collected groups', function(done) {
+  // test for 'wrapMatchGroups' method 
+  it('should count and test content of collected groups acrossElements', function(done) {
     var groupIndexes = [], groupCount = 0, content = '';
 
     new Mark($ctx[0]).markRegExp(reg, {
+      'acrossElements' : true,
       'separateGroups' : true,
       filter : function(node, group, total, obj) {
         groupIndexes.push(obj.groupIndex);
@@ -25,9 +25,11 @@ describe('test the \'collectRegexGroupIndexes\' method', function() {
         }
         return true;
       },
-      'each' : function(elem) {
-        groupCount++;
-        content += elem.textContent;
+      'each' : function(elem, info) {
+        if (info.groupStart) {
+          groupCount++;
+          content += elem.textContent;
+        }
       },
       'done' : function() {
         // 9th group is empty
@@ -39,25 +41,26 @@ describe('test the \'collectRegexGroupIndexes\' method', function() {
     });
   });
 
-  // test 'separateGroupsD' method 
-  it('should count and test content of groups with d flag2', function(done) {
+  // test for 'wrapMatchGroupsD' method
+  it('should count and test content of groups with d flag acrossElements', function(done) {
     var groupIndexes = [], groupCount = 0, content = '';
-    
     new Mark($ctx[0]).markRegExp(new RegExp(reg.source, 'd'), {
+      'acrossElements' : true,
       'separateGroups' : true,
       filter : function(node, group, total, obj) {
         groupIndexes.push(obj.groupIndex);
         // with d flag, if the parent group is filtered out, the nested group
         // will be marked - named capture group gr5
-        if (obj.groupIndex === 4
-          || obj.groupIndex === 5 || obj.groupIndex === 10) {
+        if (obj.groupIndex === 4 || obj.groupIndex === 5 || obj.groupIndex === 10) {
           return false;
         }
         return true;
       },
-      'each' : function(elem) {
-        groupCount++;
-        content += elem.textContent;
+      'each' : function(elem, info) {
+        if (info.groupStart) {
+          groupCount++;
+          content += elem.textContent;
+        }
       },
       'done' : function() {
         // 9th group is empty, the 5th nested group is added