--- a/src/lib/mark.js
+++ b/src/lib/mark.js
@@ -792,7 +792,7 @@ class Mark {
   * @param {Mark~separateGroupsEachCallback} eachCb - Each callback
   */
   separateGroups(node, match, params, filterCb, eachCb) {
-    let startIndex = 0,
+    let startIndex = match.index,
       i = -1,
       isWrapped = false,
       index, group, start;