… dist/jquery.mark.es6.js
… dist/jquery.mark.js
… dist/mark.es6.js
… dist/mark.js
… manual_test/README.md
… manual_test/manual_test_regexp_hasIndises.html
… src/lib/regexpcreator.js
… test/fixtures/across-elements/basic/block-elements-boundary.html
… test/fixtures/across-elements/basic/cache-text-nodes.html
… test/fixtures/across-elements/basic/combine-patterns.html
… test/fixtures/across-elements/basic/count.html
… test/fixtures/across-elements/basic/get-text.html
… test/fixtures/across-elements/regexp/block-elements-boundary.html
… test/fixtures/across-elements/regexp/count.html
… test/fixtures/across-elements/regexp/ignore-groups-2.html
… test/fixtures/across-elements/regexp/nested-overlapped-groups.html
… test/fixtures/across-elements/regexp/separate-groups.html
… test/fixtures/basic/cache-text-nodes.html
… test/fixtures/basic/combine-patterns.html
… test/specs/across-elements/basic/block-elements-boundary.js
… test/specs/across-elements/basic/cache-text-nodes.js
… test/specs/across-elements/basic/combine-patterns.js
… test/specs/across-elements/basic/count.js
… test/specs/across-elements/basic/done.js
… test/specs/across-elements/basic/get-text.js
… test/specs/across-elements/regexp/block-elements-boundary.js
… test/specs/across-elements/regexp/count.js
… test/specs/across-elements/regexp/separate-groups-hasIndices.js
… test/specs/across-elements/regexp/separate-groups.js
… test/specs/across-elements/regexp/wrap-nested-overlapped-groups.js