The performance test results in Firefox:
`markRegExp` method, regex `/\b(?:lorem|ipsum)\b/gi`
|                |  size 100kb, marked words 1400: |   size 200kb, marked words 1400: |   size 500kb, marked words 3000: |
|----------------|---------------------------------|----------------------------------|----------------------------------|
|  existing lib  |   ~230 ms.                      |     ~460 ms.                     |       ~2000 ms.                  |
|  PR            |   ~30 ms.                       |     ~35 ms.                      |       ~50 ms.                    |

the same without `acrossElements` option
|                |  size 100kb, marked words 1400: |   size 200kb, marked words 1400: |   size 500kb, marked words 3000: |
|----------------|---------------------------------|----------------------------------|----------------------------------|
|  existing lib  |   ~50 ms.                       |     ~70 ms.                      |       ~120 ms.                   |
|  PR            |   ~30 ms.                       |     ~45 ms.                      |       ~60 ms.                    |

`markRanges` method - generated number of ranges equal marked word number
|                |  size 100kb, marked words 1400: |   size 200kb, marked words 1400: |   size 500kb, marked words 3000: |
|----------------|---------------------------------|----------------------------------|----------------------------------|
| existing lib   |   ~80 ms.                       |     ~140 ms.                     |       ~620 ms.                   |
| PR             |   ~30 ms.                       |     ~35 ms.                      |       ~70 ms.                    |

Also see below <a href="#user-content-ways-to-boost-performance">how to boost performance in `mark()` method.</a>  

The 8MB file containing 177000 text nodes, marked words 5000:
in Firefox  ~470 ms.
in Chrome ~1300 ms.
The big difference in the performance of Firefox vs. Chrome on my computer is probably related to the processor cache.
For 3.2MB file containing 69200 text nodes and 2400 marked words is only 180 ms. vs 200 ms.

Before modification in the 'iterateThroughNodes' method performance was:
in Firefox  ~1150 ms.
in Chrome ~3600 ms.

The `markRegExp` method, with the `separateGroups` option, now has the ability to mark separate groups.  
With `wrapAllRanges` option can mark nesting and overlapping ranges and match groups.

To the callbacks `each` and `filter` in the `markRegExp` and `mark` methods i added the additional parameters - matchInfo and filterInfo objects containing match information.
The additional parameters allow to use of simple code to correctly count words and/or phrases, pinpoint the start of matches, which are located across elements, abort execution, match.input property exposes internal string ...

They will not break backward-compatibility but can help to solve many problems without 'happy hacking'.

`filterInfo`  object in the code examples:
``` js
/**
* @property {array} match - The result of RegExp exec() method
* @property {boolean} matchStart - indicate the start of match
* @property {number} groupIndex - The group index, is only available with 'separateGroups' options
* @property {object} execution - The helper object for early abort. Contains boolean 'abort' property.
* @property {number} offset - The absolute start index of a text node.
* Can be used to translate the local node indexes to the absolute ones
*/
 ```
`info` or `matchInfo` object in the code examples:
``` js
/**
* @property {array} match - The result of RegExp exec() method
* @property {boolean} matchStart - indicate the start of match
* @property {number} count - The current number of matches
* @property {number} groupIndex - The index of match group, is only available with 'separateGroups' options
* @property {boolean} groupStart - indicate the start of group, is only
* available with both 'acrossElements' and 'separateGroups' options
*/
```
To the `done` callbacks in the `mark`, `markRegExp`, and `markRanges` methods I added the exact total match counter, and `termStats` object in the `mark` method.
The `done` callback in the code examples:
``` js
/**
* @param {number} totalMarks - The total number of marked elements
* @param {number} totalMatches - The exact number of total matches
* @param {object} termStats - An object containing an individual term's matches count for `mark` method
*/
```

Available properties of the `filterInfo` object on the `filter` callback depending on options
mark - `filter : function(textNode, foundTerm, totalMarks, counter, filterInfo) {}`
markRegExp - `filter : function(textNode, foundTerm, totalMarks, filterInfo) {}`

|  method      |            options               |     match      |   matchStart    |   groupIndex  |  execution  | offset |
|--------------|----------------------------------|----------------|-----------------|---------------|-------------|--------|
|  mark        |  acrossElements                  |      +         |      +          |      -        |     +       |   +    |
|  mark        |                                  |      +         |      -          |      -        |     +       |   +    |
|              |                                  |                |                 |               |             |        |
|  markRegExp  |  acrossElements                  |      +         |      +          |      -        |     +       |   +    |
|  markRegExp  |  acrossElements, separateGroups  |      +         |      +          |      +        |     +       |   -    |
|  markRegExp  |  separateGroups                  |      +         |      +          |      +        |     +       |   +    |
|  markRegExp  |                                  |      +         |      -          |      -        |     +       |   -    |


Available properties of the `matchInfo` object on the `each` callback depending on options
mark, markRegExp - `each : function(node, matchInfo) {}`

|  method      |             options              |      match     |    matchStart   |  groupIndex  | groupStart | count |
|--------------|----------------------------------|----------------|-----------------|--------------|------------|-------|
|  mark        |  acrossElements                  |       +        |      +          |     -        |     -      |   +   |
|  mark        |                                  |       +        |      -          |     -        |     -      |   +   |
|              |                                  |                |                 |              |            |   +   |
|  markRegExp  |  acrossElements                  |       +        |      +          |     -        |     -      |   +   |
|  markRegExp  |  acrossElements, separateGroups  |       +        |      +          |     +        |     +      |   +   |
|  markRegExp  |  separateGroups                  |       +        |      +          |     +        |     -      |   +   |
|  markRegExp  |                                  |       +        |      -          |     -        |     -      |   +   |


The start mark elements in the code examples are only useful with 'acrossElements' option in cases such as with next/previous buttons to highlight match no matter how many mark elements it contains. The `data-markjs` attribute can be used for this purpose instead.

The `markRegExp()` method code example with `acrossElements` option
``` js
let matchCount = 0;

context.markRegExp(/AB\s+BC\s+EF/gi, {
    'acrossElements' : true,
    'each' : function(elem, info) {
        // for external counter 
        matchCount = info.count;
        
        // for internal use
        if(info.count ..) {}
        
        // if start of the match
        if(info.matchStart) {
            elem.className = 'start-1';
            // also possible
            // matchCount++;
        }
        
        // info.count as a match identifier
        // elem.setAttribute('data-markjs', info.count);
    },
    'done' : function(totalMarks, totalMatches) {
        console.log('Total matches = ' + totalMatches);
    }
});
```
The `mark()` method code example with `acrossElements` option 
``` js
let matchCount = 0;

context.mark(['AB CD', 'EF'], {
    'separateWordSearch' : true,
    'acrossElements' : true,
    'each' : function(elem, info) {
       // the counter usage is the same as in markRegExp() method above
        
    },
    'done' : function(totalMarks, totalMatches, termStats) {
        console.log('Total matches = ' + totalMatches);
        
        for(var term in termStats) {
            console.log(term + ' = ' + termStats[term]);
        }
    }
});
```
The `mark()` method code example without `acrossElements` option 
``` js
let matchCount = 0;

context.mark('AB CD EF', {
    'separateWordSearch' : true,
    'each' : function(elem, info) {
        // for external counter
        matchCount = info.count; // also possible matchCount++;
        
        // for internal use
        if(info.count ..) {}
    },
    'done' : function(totalMarks, totalMatches, termStats) {
        // Note: although without 'acrossElements' option 'totalMarks' and 'totalMatches' are equal,
        // the latter should be preferred
        console.log('Total matches = ' + totalMatches);
        
        for(var term in termStats) {
            console.log(term + ' = ' + termStats[term]);
        }
    }
});
```
The `markRanges()` method code example
``` js
let ranges = [{ start: 15, length: 50 }, {..}, {..}];

context.markRanges(ranges, {
    'done' : function(totalMarks, totalMatches) {
        console.log('Total range matches = ' + totalMatches);
    }
});
```

<h2 id="filter-matches">Filter matches</h2>

Filter matches in the `markRegExp()` method
``` js
let count = 0, reg = /\bAB\b.+?\bCD\b/gi;

// if you have access to the RegExp object with 'acrossElements' option, you can also
// set 'lastIndex' to 'Infinity' to break the execution
context.markRegExp(reg, {
    filter : function(node, matchStr, totalMarks, filterInfo) {
        // to mark only the first match
        filterInfo.execution.abort = true; return  true;
        // reg.lastIndex = Infinity; return  true; // only with 'acrossElements' option

        // filter callback requires its own match counter
        if(filterInfo.matchStart) {
            count++;
        }
        // mark the first 10 matches.
        if(count > 10) {
            filterInfo.execution.abort = true;
            // reg.lastIndex = Infinity; // only with 'acrossElements' option
            return  false;
        }

        // skip between
        if(count > 10 && count < 20) { return  false; }

        // mark between
        if(count <= 10) { return  false; }
        else if(count > 20) {
            filterInfo.execution.abort = true;
            // reg.lastIndex = Infinity; // only with 'acrossElements' option
            return  false;
        }

        return  true;
    },
});

```
Mark the first desired number of matches on `each` callback with `acrossElements` option. It's much limited than the `filter` callback.
``` js
let reg = /\b(A\s+B\s+C)\b/gi;

context.markRegExp(reg, {
    'acrossElements' : true,
    'each' : function(elem, info) {
        // to mark only the first match
        reg.lastIndex = Infinity;

        // first 10 matches
        if(info.count >= 10) {
            reg.lastIndex = Infinity;
        }
    }
});
```
Filter matches in the `mark` method with `acrossElements` option
``` js
let count = 0;

context.mark('AB', {
    'acrossElements' : true,
    filter : function(node, term, totalMarks, currentMarks, filterInfo) {
         // to mark only the first match
        filterInfo.execution.abort = true; return  true;

        // filter callback requires its own match counter
        if(filterInfo.matchStart) {
            count++;
        }
        // mark the first 10 matches.
        if(count > 10) { filterInfo.execution.abort = true; return  false; }

        // skip between
        if(count > 10 && count < 20) { return  false; }

        // mark between
        if(count <= 10) { return  false; }
        else if(count > 20) { filterInfo.execution.abort = true; return  false; }

        return  true;
    }
});
```
Filter matches in the `mark` method without `acrossElements` option
``` js
let count = 0;

context.mark('AB', {
    filter : function(node, term, totalMarks, currentMarks, filterInfo) {
        // the only difference is counter implementation
        count++;
    }
});
```

<h2 id="mark-separate-groups">Mark separate groups</h2>

**Important:** in this implementation two branches of code process separate groups, which one, depending on the existence of `d` flag.
1. Primitive, base on `indexOf()`, only reliable with contiguous groups - unwanted group(s) can be easily filtered out.
2. Exact, but not all browsers currently supported group `indices`.

They both have identical logic for nested groups - if the parent group has been marked, there is no way to mark nested groups.
This means you can use nested group(s) as auxiliary and don't care about filtering them.  
Note: the new `wrapAllRanges` option allows change this behavior.

But they have different parent groups logic:
* The exact one does allow to use parent group as auxiliary - you need to filter out it in order to mark the nested group(s).
* The primitive one does not allow this - if the parent group has filtered out the nested group(s) won't be marked.

Note: Code which will use the primitive implementation will be compatible with the exact if the unwanted group(s) don't contain nested group(s) (in the exact implementation, if parent group is filtered out, the nested group(s) will be marked).
To test the primitive one just add `d` flag.

Although, there is no strict requirement for the contiguity of capture groups.
compare:
in string - 'AAB xxx BCD xx BC', to mark groups AB and BC.
in /(AB)\b.+?\b(BC)/g the indexOf('BC', start) find first 'BC', which is correct,
but in /(AB)\b(.+?)\b(BC)(?!D)/g the indexOf('BC', start) also find first 'BC', which is wrong, because of condition '(?!D)', so group 2 is required.

Warning: related using RegExp without `d` flag:
* Do not add the capture group(s) to lookbehind assertion `(?<=)`, there is no code which handle such cases.
* With `acrossElements` option, currently not possible to highlight the capture group(s) inside the lookahead assertion (?=).

Example to mark separate groups with `acrossElements` option
``` js
let groupCount = 0, group1Count = 0, group2Count = 0;

context.markRegExp(/\b(AB)\b.+?\b(CD)\b/gi, {
    'acrossElements' : true,
    'separateGroups' : true,
    'each' : function(elem, info) {
        // if start of match group
        if(info.groupStart) {
            // all group count
            groupCount++;
            
            // info.groupIndex is index of the current match group
            if(info.groupIndex === 1) {
                elem.className = 'group1-1';
                // individual group count
                group1Count++;

            } else if(info.groupIndex === 2) {
                elem.className = 'group2-1';
                group2Count++;
            }
        }
    }
});
```
Example to mark separate groups without `acrossElements` option
``` js
let groupCount = 0, group1Count = 0, group2Count = 0;

context.markRegExp(/\b(AB)\b.+?\b(CD)\b/gi, {
    'separateGroups' : true,
    'each' : function(elem, info) {
        // all group count
        groupCount++;
        
        // info.groupIndex is index of the current match group
        if(info.groupIndex === 1) {
            // individual group count
            group1Count++;

        } else if(info.groupIndex === 2) {
            group2Count++;
        }
    }
});
```
Filter capture groups  - if filter return false, the group will be ignored
``` js
context.markRegExp(/(AB)\b(.+)\b(CD)(...)(?<gr5>EF)?\b/gi, {
    // with and without `acrossElements` option
    'acrossElements' : true,
    'separateGroups' : true,
    filter : function(node, group, totalMarks, filterInfo) {
        // by group index
        // filterInfo.groupIndex - current group index. Note: if group lays across several elements,
        // the index will be the same while the current group is wrapping
        if(filterInfo.groupIndex === 2 || filterInfo.groupIndex === 4) return  false;

        // by group content
       // if(filterInfo.group === 'AB') return  false;

        // to filter the whole match. Note: it will iterate through all groups and only then return
        if(filterInfo.match[5]) return true/false;

        // Note: named capture groups can be use only to filter whole match
        if(filterInfo.match.groups.gr5) return  true/false;

        return  true;
    },
});
```

<h2 id="elements-boundaries">Elements boundaries</h2>

With the `acrossElements` option, text nodes are aggregated into a single string, taking into account Html elements.
If the block element divides text nodes, and the first text node doesn't end by white space, the space is added to the string to separate them.

With the additional `blockElementsBoundary` option, if the text node doesn't end by white space - ' \u001 ', otherwise - '\u001 ' string is added between them.
It's become something 'across inline elements' option.

The `blockElementsBoundary` option only make sense when highlighting phrases or RegExp separate groups.
It will keep the matches from crossing the boundary of block elements or only custom elements.

``` js
context.markRegExp(/\bAB\s+BC\s+CD\b/gi, {
    'acrossElements' : true,
    // option divide text nodes by ' \u001 ' instead of space
    // 'blockElementsBoundary' : {}, or with options
    'blockElementsBoundary' : {
        // custom block elements - only this custom elements will have boundaries
        'tagNames' : ['div', 'p', 'h1', 'h2'],    //optional
        // custom boundary char, default is '\u001'
        'char' : '.'        //optional
    }
});
```

<h2 id="mark-nesting-overlapping">Mark nesting and overlapping ranges and match groups</h2>

The `markRanges()` method with `wrapAllRanges` option, can mark nesting/overlapping ranges.

The `markRegExp()` method with RegExp having the `d` flag, with `separateGroups` and `wrapAllRanges` options can mark:
nesting groups, capturing groups inside positive lookaround assertions. It's practically removes all restrictions.  

The lookaround examples below demonstrate cases when `wrapAllRanges` option should be used, otherwise it won't be correctly highlighted:

* RegExp with lookaround assertions can create overlapping matches.  
  e.g. regex `/(?<=(gr1)\s+\w+\b).+?(gr2)/dg`,  string 'gr1 match1 gr1 gr2 match2 gr2'.  
  The gr1 from the second match won't be wrap because the gr2 from the first match is already wrapped.

* Another case: regex `/(?=\d*(1))(?=\d*(2))(?=\d*(3))/dg`, matches '123, 132, 213, 231, 312, 321'.  
  This is not an overlapping case, but groups are wrapped in any order. If group 1 is wrapped first, the 2 and 3 will be ignored in '231, 321' ...  

* Groups overlapping case: regex `/\w+(?=.*?(gr1 \w+))(?=.*?(\w+ gr2))/dg` , string 'word gr1 overlap gr2' - the gr1 will be wrapped, the gr2 will be ignored.

Note: the `wrapAllRanges` option can cause performance degradation if the context contains a very large number of text nodes and a large number of mark elements. 
This is because with each wrapping, two more objects are inserted into the array, which require a lot of copying, memory allocation ...

The 8MB file containing 177000 text nodes:
…