@dabrahams 2021-09-28
There is a backward-compatibility issue I noticed when I dropped this change into my project: it now insists on `g` or `y` flags for the regex in a regex search.