@dabrahams 2021-09-28
I think this patch needs some documentation for `matchNodeIndex` (and anything else added).  I don't know what it means and all the examples merely test it against zero, so it's hard to guess.