@angezid 2021-09-29
> There is a backward-compatibility issue I noticed when I dropped this change into my project: it now insists on g or y flags for the regex

The first idea was to silently recompile RegExp with the `g` flag, but how to control this new RegExp? Without `g` or `y` flags there will be an infinite loop.
In the existing library an infinite loop is control by complex manipulations with input string and nodes indexes.