@dabrahams 2021-09-29
Can't you just recompile it with `g` if and only if the object has neither `.global == true` nor `.sticky == true`?