@angezid 2021-09-30
In my previous versions there was code which do this: 
``` js
if (this.opt.acrossElements && !regexp.global && !regexp.sticky) {
  let flags = 'g' + (regexp.flags ? regexp.flags : '');
  regexp = new RegExp(regexp.source, flags);
}
```
but what if user forget to set `g` flag and later try `regexp.lastIndex = Number.MAX_VALUE;`?