@angezid 2021-12-23
@julmot, I implement what I want, even more than. 
But how to test `wrapMatchGroupsD` and `separateGroupsD` methods?
With PhantomJS it's not possible to test even RegExp pattern parser?
Currently, I left the manual test.
Also, some doubts about correct naming, comments ...