@drzraf 2022-06-04
I'm trying to anticipate `mark.js` future and that led me to wonder about maintainer's vision (@julmot) regarding this (remarkable) PR (and the overall project).