@drzraf 2022-06-06
> To the callbacks each and filter in the markRegExp and mark methods i added the additional parameters - matchInfo and filterInfo objects containing match information.

This is great. When using `markRegExp` with a regexp containing named groups, these would be available inside `matchInfo.groups`. Very useful.