@rbhalla 2022-06-13
@angezid, I was looking at contributing the following enhancement to the existing repo:
https://github.com/julmot/mark.js/issues/473

Then I found this incredible PR. Would you open to contributions to this branch? It seems well updated, and could easily form the basis of a fork.

Not sure what you think of the issue, but it feels like `acrossElements: true` and `separateWordSearch: false` is a useful combination to support.