@drzraf 2022-07-14
Just wanted to mention that I had a web-extension. I couldn't rely upon `mark.js` because of the low performance, but using this branch saved my day and made it just work like a breeze.
Big thanks to @angezid !