@stevemckenzie 2022-07-28
@angezid you should reach out to @julmot 🙃  - https://github.com/julmot/mark.js/issues/463#issuecomment-1048890093