@angezid 2022-07-29
He is probably very busy. Version 9.0.0 still not released [https://github.com/julmot/mark.js/issues/457](https://github.com/julmot/mark.js/issues/457).