@stevemckenzie 2022-08-02
@angezid I reached out to him via email. He is definitely too busy to maintain this so maybe you can help to get version 9 out the door? He said he was waiting for you to respond.