@peace2000 2022-08-05
If this is going through changes, would it be possible to add support for texts under shadow root? https://github.com/julmot/mark.js/issues/458