@stevemckenzie 2022-08-16
@angezid did you connect with @julmot yet?