@julmot 2022-08-16
@stevemckenzie No, I've also sent a reminder comment [here](https://github.com/julmot/mark.js/issues/481#issuecomment-1203803899).